Backend connected

+++ springapp/src/main/java/com/examly/springapp/repository/HealthCoachRepo.java
@@ -3,4 +3,5 @@ package com.examly.springapp.repository;
 import com.examly.springapp.model.HealthCoach;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface HealthCoachRepo extends JpaRepository<HealthCoach, Integer> {}
+public interface HealthCoachRepo extends JpaRepository<HealthCoach, Long> {
+}
